refactor: input area types
