+++ projects/ion/src/lib/core/types/input-area.ts
@@ -1,0 +1,10 @@
+import { EventEmitter } from '@angular/core';
+
+export interface IonInputAreaProps {
+  cols?: string;
+  rows?: string;
+  disabled?: boolean;
+  value?: string;
+  placeholder?: string;
+  valueChange?: EventEmitter<string>;
+}
