+++ projects/ion/src/lib/input-area/input-area.component.ts
@@ -1,14 +1,5 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 
-export interface IonInputAreaProps {
-  cols?: string;
-  rows?: string;
-  disabled?: boolean;
-  value?: string;
-  placeholder?: string;
-  valueChange?: EventEmitter<string>;
-}
-
 @Component({
   selector: 'ion-input-area',
   templateUrl: './input-area.component.html',
